pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlflow-enterprise-gateway"
version = "0.1.0"
description = "Policy Enforcement Gateway extension layer for MLflow multi-tenancy, RBAC context, and audit logging"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "PyJWT[crypto]>=2.9.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "respx>=0.22.0"
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools]
packages = ["gateway", "gateway.mlflow"]
